Title: 시력 검사 2, Time: 130ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240202/시력 검사 2/eye-test-2.py
@@ -1,4 +1,4 @@
-dpt=int(input())
+dpt=float(input())
 if dpt>=1.0:print('High')
 elif dpt>=0.5:print('Middle')
 else:print('Low')
